Consolidated all user endpoints into once class
